# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bankcards API
  version: 1.0.0

servers:
  - url: /api/v1

paths:
  /registration:
    post:
      tags:
        - Регистрация
      summary: "Регистрация"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: "OK"
        '400':
          description: "Пользователь с таким именем уже зарегистрирован"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth:
    post:
      tags:
        - Аутентификация
      summary: Аутентификация пользователя
      requestBody:
        description: Учетные данные пользователя
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  
  /auth/refresh-token:
    post:
      tags:
        - Аутентификация
      summary: Обновление токена доступа
      description: Необходимо отправить refresh-token в заголовке Authorization
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Токены успешно обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Недействительный или просроченный Refresh Token
        '500':
          description: Внутренняя ошибка сервера
  /cards:
    post:
      summary: "Создать карту"
      tags:
        - Управление картами
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
        "403":
          description: "FORBIDDEN"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: "Вывести все карты"
      tags:
        - Управление картами
      parameters:
        - name: "page"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int32"
        - name: "size"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CardDTO"
        "403":
          description: "FORBIDDEN"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cards/change-status:
    patch:
      summary: "Изменить статус карты"
      tags:
        - Управление картами
      parameters:
        - name: "cardId"
          in: "query"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "newStatus"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - "ACTIVE"
              - "BLOCKED"
              - "EXPIRED"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
        "403":
          description: "FORBIDDEN"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cards/{cardId}:
    delete:
      summary: "Удалить карту по ID"
      tags:
        - Управление картами
      parameters:
        - name: "cardId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
        "403":
          description: "FORBIDDEN"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        password:
          type: string
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: '400'
        message:
          type: string
        timestamp:
          type: string
    CreateCardDTO:
      type: "object"
      properties:
        pan:
          type: "string"
        balance:
          type: string
          format: decimal
          example: "1500.50"
        ownerId:
          type: "string"
          format: "uuid"
        expiredAt:
          type: "string"
          format: "date"
    CardDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        maskedPan:
          type: "string"
        status:
          type: "string"
          enum:
            - "ACTIVE"
            - "BLOCKED"
            - "EXPIRED"
        balance:
          type: string
          format: decimal
          example: "1500.50"
        expiredAt:
          type: "string"
          format: "date"
        ownerId:
          type: "string"
          format: "uuid"
security:
  - bearerAuth: []